TravelTimeGoogleApp moved to utils; return type declarations.
